organized myevents card into its own component
