+++ code/frontend/src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useState } from "react";
-import { Link } from "react-router-dom";
 import EventCalendar from "../components/EventCalendar";
 import EventCard from "../components/EventCard";
+import MyEvents from "../components/MyEvents";
 import { supabase } from "../lib/supabase";
 import "../styles/Home.css";
 
@@ -139,13 +139,6 @@ const Home = ({ session }) => {
   const [categories, setCategories] = useState([ALL_EVENTS_CATEGORY]);
   const [eventsLoading, setEventsLoading] = useState(true);
   const [eventsError, setEventsError] = useState("");
-  const [myEvents, setMyEvents] = useState([]);
-  const [myEventsLoading, setMyEventsLoading] = useState(false);
-  const [myEventsError, setMyEventsError] = useState("");
-  const [myEventsSections, setMyEventsSections] = useState({
-    pending: true,
-    approved: true,
-  });
   const [selectedDate, setSelectedDate] = useState(null);
   const [selectedCategory, setSelectedCategory] = useState(ALL_EVENTS_CATEGORY);
   const [searchQuery, setSearchQuery] = useState("");
@@ -240,59 +233,6 @@ const Home = ({ session }) => {
     };
   }, []);
 
-  useEffect(() => {
-    let isMounted = true;
-
-    const fetchMyEvents = async () => {
-      if (!session?.user?.id) {
-        setMyEvents([]);
-        setMyEventsError("");
-        setMyEventsLoading(false);
-        return;
-      }
-
-      setMyEventsLoading(true);
-      setMyEventsError("");
-
-      const { data, error } = await supabase
-        .from("events")
-        .select(
-          `
-            id,
-            title,
-            start_datetime,
-            end_datetime,
-            status,
-            created_by,
-            organizations ( name )
-          `
-        )
-        .eq("created_by", session.user.id)
-        .order("start_datetime", { ascending: true });
-
-      if (!isMounted) {
-        return;
-      }
-
-      if (error) {
-        console.error("Error fetching user events:", error);
-        setMyEvents([]);
-        setMyEventsError("We couldn't load your submitted events right now.");
-        setMyEventsLoading(false);
-        return;
-      }
-
-      setMyEvents((data || []).map(normalizeSupabaseEvent));
-      setMyEventsLoading(false);
-    };
-
-    fetchMyEvents();
-
-    return () => {
-      isMounted = false;
-    };
-  }, [session]);
-
   const calendarDays = useMemo(
     () => getMonthMatrix(visibleMonth.getFullYear(), visibleMonth.getMonth()),
     [visibleMonth]
@@ -327,52 +267,6 @@ const Home = ({ session }) => {
     });
   }, [events, searchQuery, selectedCategory, selectedDate]);
 
-  const myEventCounts = useMemo(() => {
-    return myEvents.reduce(
-      (counts, event) => {
-        const statusKey = event.status?.toLowerCase() || "other";
-
-        if (statusKey === "approved") {
-          counts.approved += 1;
-        } else if (statusKey === "pending") {
-          counts.pending += 1;
-        } else {
-          counts.other += 1;
-        }
-
-        return counts;
-      },
-      { approved: 0, pending: 0, other: 0 }
-    );
-  }, [myEvents]);
-
-  const pendingMyEvents = useMemo(
-    () => myEvents.filter((event) => (event.status || "").toLowerCase() === "pending"),
-    [myEvents]
-  );
-
-  const approvedMyEvents = useMemo(
-    () => myEvents.filter((event) => (event.status || "").toLowerCase() === "approved"),
-    [myEvents]
-  );
-
-  const otherMyEvents = useMemo(
-    () =>
-      myEvents.filter((event) => {
-        const statusKey = (event.status || "").toLowerCase();
-        return statusKey !== "pending" && statusKey !== "approved";
-      }),
-    [myEvents]
-  );
-
-  function getStatusLabel(status) {
-    if (!status) {
-      return "In Review";
-    }
-
-    return status.charAt(0).toUpperCase() + status.slice(1);
-  }
-
   const monthLabel = visibleMonth.toLocaleDateString("en-US", {
     month: "long",
     year: "numeric",
@@ -394,13 +288,6 @@ const Home = ({ session }) => {
 
   function resetDateFilter() {
     setSelectedDate(null);
-  }
-
-  function toggleMyEventsSection(sectionName) {
-    setMyEventsSections((currentSections) => ({
-      ...currentSections,
-      [sectionName]: !currentSections[sectionName],
-    }));
   }
 
   return (
@@ -518,144 +405,11 @@ const Home = ({ session }) => {
           </section>
 
           {session ? (
-            <aside className="home-right-column">
-              <div className="home-panel home-my-events-panel">
-                <div className="panel-header">
-                  <h2>My Events</h2>
-                  <p className="panel-description">
-                    Track where your community submissions stand.
-                  </p>
-                </div>
-
-                <div className="my-events-summary">
-                  <div className="my-events-stat">
-                    <span className="my-events-stat-label">Pending</span>
-                    <strong>{myEventCounts.pending}</strong>
-                  </div>
-                  <div className="my-events-stat">
-                    <span className="my-events-stat-label">Approved</span>
-                    <strong>{myEventCounts.approved}</strong>
-                  </div>
-                </div>
-
-                {myEventsLoading ? (
-                  <div className="my-events-empty-state">
-                    <h3>Loading Your Events</h3>
-                    <p>Checking the status of your recent submissions.</p>
-                  </div>
-                ) : myEventsError ? (
-                  <div className="my-events-empty-state">
-                    <h3>Unable to Load</h3>
-                    <p>{myEventsError}</p>
-                  </div>
-                ) : myEvents.length === 0 ? (
-                  <div className="my-events-empty-state">
-                    <h3>No Submitted Events Yet</h3>
-                    <p>
-                      When you submit an event for review, it will appear here with its
-                      approval status.
-                    </p>
-                    <Link className="my-events-cta" to="/post-event">
-                      Post an Event
-                    </Link>
-                  </div>
-                ) : (
-                  <div className="my-events-list">
-                    <section className="my-events-group">
-                      <button
-                        className="my-events-group-toggle"
-                        onClick={() => toggleMyEventsSection("pending")}
-                        type="button"
-                      >
-                        <span className="my-events-group-title">Pending Events</span>
-                        <span className="my-events-group-meta">
-                          {pendingMyEvents.length}
-                          <span className="my-events-group-chevron">
-                            {myEventsSections.pending ? "−" : "+"}
-                          </span>
-                        </span>
-                      </button>
-
-                      {myEventsSections.pending ? (
-                        pendingMyEvents.length > 0 ? (
-                          <div className="my-events-group-list">
-                            {pendingMyEvents.map((event) => (
-                              <article className="my-event-card" key={event.id}>
-                                <div className="my-event-card-top">
-                                  <span className="my-event-status my-event-status-pending">
-                                    {getStatusLabel(event.status)}
-                                  </span>
-                                  <span className="my-event-date">
-                                    {formatCompactDate(event.startDate)}
-                                  </span>
-                                </div>
-                                <h3>{event.title}</h3>
-                                <p className="my-event-org">{event.organizationName}</p>
-                                <p className="my-event-time">
-                                  {formatTimeRange(event.startDate, event.endDate)}
-                                </p>
-                              </article>
-                            ))}
-                          </div>
-                        ) : (
-                          <p className="my-events-group-empty">
-                            No pending events right now.
-                          </p>
-                        )
-                      ) : null}
-                    </section>
-
-                    <section className="my-events-group">
-                      <button
-                        className="my-events-group-toggle"
-                        onClick={() => toggleMyEventsSection("approved")}
-                        type="button"
-                      >
-                        <span className="my-events-group-title">Approved Events</span>
-                        <span className="my-events-group-meta">
-                          {approvedMyEvents.length + otherMyEvents.length}
-                          <span className="my-events-group-chevron">
-                            {myEventsSections.approved ? "−" : "+"}
-                          </span>
-                        </span>
-                      </button>
-
-                      {myEventsSections.approved ? (
-                        approvedMyEvents.length > 0 || otherMyEvents.length > 0 ? (
-                          <div className="my-events-group-list">
-                            {[...approvedMyEvents, ...otherMyEvents].map((event) => (
-                              <article className="my-event-card" key={event.id}>
-                                <div className="my-event-card-top">
-                                  <span
-                                    className={`my-event-status my-event-status-${(
-                                      event.status || "approved"
-                                    ).toLowerCase()}`}
-                                  >
-                                    {getStatusLabel(event.status)}
-                                  </span>
-                                  <span className="my-event-date">
-                                    {formatCompactDate(event.startDate)}
-                                  </span>
-                                </div>
-                                <h3>{event.title}</h3>
-                                <p className="my-event-org">{event.organizationName}</p>
-                                <p className="my-event-time">
-                                  {formatTimeRange(event.startDate, event.endDate)}
-                                </p>
-                              </article>
-                            ))}
-                          </div>
-                        ) : (
-                          <p className="my-events-group-empty">
-                            No approved events yet.
-                          </p>
-                        )
-                      ) : null}
-                    </section>
-                  </div>
-                )}
-              </div>
-            </aside>
+            <MyEvents
+              session={session}
+              formatCompactDate={formatCompactDate}
+              formatTimeRange={formatTimeRange}
+            />
           ) : null}
         </section>
       </main>
